restructure test scripts
